frontend/requirements.txt
flask
flask-restx
flask_cors
pymongo
flask_cors
gunicorn
python-dotenv